+++ python/selfie-lib/selfie_lib/__init__.py
@@ -1,17 +1,19 @@
 # maintain alphabetical order
+from .ArrayMap import ArrayMap as ArrayMap
+from .CommentTracker import CommentTracker as CommentTracker
+from .FS import FS as FS
 from .LineReader import LineReader as LineReader
 from .Literals import LiteralValue as LiteralValue
-from .Mode import Mode as Mode
 from .ParseException import ParseException as ParseException
 from .PerCharacterEscaper import PerCharacterEscaper as PerCharacterEscaper
 from .Slice import Slice as Slice
 from .Snapshot import Snapshot as Snapshot
+from .SnapshotFile import SnapshotFile as SnapshotFile
 from .SnapshotReader import SnapshotReader as SnapshotReader
 from .SnapshotSystem import _initSelfieSystem as _initSelfieSystem
 from .SnapshotSystem import _selfieSystem as _selfieSystem
 from .SnapshotSystem import DiskStorage as DiskStorage
-from .SnapshotSystem import FS as FS
-from .SnapshotSystem import SnapshotFile as SnapshotFile
+from .SnapshotSystem import Mode as Mode
 from .SnapshotSystem import SnapshotSystem as SnapshotSystem
 from .SnapshotValue import SnapshotValue as SnapshotValue
 from .SnapshotValueReader import SnapshotValueReader as SnapshotValueReader
